refactor: Extract `getTemplateAfterAspectInvocation` for more usage
