+++ src/utils/test/index.ts
@@ -1,3 +1,4 @@
 export * from "./assertions";
 export * from "./simple-gu-stack";
 export * from "./attach-policy-to-test-role";
+export * from "./template";
